Split direct build file-read host effect fixtures
