+++ tests/integration/cli_build/direct_build_graph_host_effects/file_reads/mod.rs
@@ -1,0 +1,2 @@
+mod declared;
+mod rejections;
